requirements.txt
streamlit
gspread
pandas
plotly
oauth2client
Pillow